gemini-enterprise/custom-connector/requirements.txt
google-auth
google-auth-httplib2
google-cloud-discoveryengine
requests
watchfiles
